refactor:removed Chat-DevSetup module
