+++ Chat-Backend/.gitignore
@@ -4,3 +4,5 @@ bin
 .project
 .classpath
 /target/
+.idea/
+*.iml
