+++ Chat-Backend/src/main/java/com/github/sourcecase/chat/ChatWebAppConfiguration.java
@@ -8,9 +8,6 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
 public class ChatWebAppConfiguration extends WebMvcConfigurerAdapter {
 
 	public static void main(String[] args) throws Exception {
-//		if (AuthConfigFactory.getFactory() == null) {
-//			AuthConfigFactory.setFactory(new AuthConfigFactoryImpl());
-//		}
 		SpringApplication.run(ChatWebAppConfiguration.class, args);
 	}
