+++ Chat-Backend/src/main/java/com/github/sourcecase/chat/web/config/ChatPathConfiguration.java
@@ -1,4 +1,4 @@
-package com.github.sourcecase.chat.web;
+package com.github.sourcecase.chat.web.config;
 
 public final class ChatPathConfiguration {
