--- Chat-DevSetup/cmd_git.bat
@@ -1,2 +1,0 @@
-cd D:\Development\chat-app\
-START cmd.exe
