--- Chat-DevSetup/cmd_jetty.bat
@@ -1,2 +1,0 @@
-cd D:\Programme\jetty-distribution-9.3.7.v20160115\demo-base
-java -jar ../start.jar
