--- Chat-DevSetup/cmd_root.bat
@@ -1,2 +1,0 @@
-cd D:\Development\chat-app\Chat-Backend
-START cmd.exe
