+++ README.md
@@ -1,9 +1,17 @@
 # chat-app
 a basic spring based chat application
 
-# usage
-install maven from https://maven.apache.org/  
-git clone https://github.com/sourcecase/chat-app.git  
-cd ${chat-app}/Chat-Backend  
-mvn jetty:run  
-go to http://localhost:8080/chat/
+# start
+
+    install maven from https://maven.apache.org/
+    git clone https://github.com/sourcecase/chat-app.git
+    cd chat-app/Chat-Backend
+    mvn install
+    java -jar Chat-Backend/target/Chat-Backend-x-y-z.RELEASE.jar
+    go to http://localhost:8080/
+    
+# use
+
+    Login with "Chris" and "asdf" as password.
+    Login with another browser as "Alice" and "asdf" and start chatting.
+
