--- pom.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.4.2.RELEASE</version>
-	</parent>
-
-	<groupId>com.github.sourcecase.chat</groupId>
-	<artifactId>Chat-Root</artifactId>
-	<name>Chat-Root</name>
-	<packaging>pom</packaging>
-	<description>A simple spring based chat system</description>
-	<version>0.3.0-SNAPSHOT</version>
-
-	<properties>
-		<java-version>1.8</java-version>
-		<org.springframework.version>4.2.5.RELEASE</org.springframework.version>
-		<org.springframework.security.version>4.0.4.RELEASE</org.springframework.security.version>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-	</properties>
-
-	<modules>
-		<module>Chat-Backend</module>
-	</modules>
-
-	<dependencyManagement>
-		<dependencies>
-
-			<!-- modules -->
-			<dependency>
-				<groupId>com.github.sourcecase.chat.backend</groupId>
-				<artifactId>chat-spring-backend</artifactId>
-				<version>0.3.0-SNAPSHOT</version>
-			</dependency>
-
-		</dependencies>
-	</dependencyManagement>
-
-	<build>
-		<plugins>
-
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-release-plugin</artifactId>
-				<version>2.5.3</version>
-			</plugin>
-
-		</plugins>
-
-		<pluginManagement>
-			<plugins>
-
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-compiler-plugin</artifactId>
-					<version>3.5.1</version>
-					<configuration>
-						<source>${java-version}</source>
-						<target>${java-version}</target>
-						<debug>true</debug>
-						<compilerArgument>-Xlint:all</compilerArgument>
-						<showWarnings>true</showWarnings>
-						<showDeprecation>true</showDeprecation>
-					</configuration>
-				</plugin>
-
-			</plugins>
-		</pluginManagement>
-
-	</build>
-
-</project>
